Implement client whitelist for TLS
A client which enters the correct secret to UDP is added to
a whitelist. Only whitelisted client IP addresses
are allowed to initiate a TCP/TLS connection.

The whitelist is saved as global struct, and thus reset between runs.
This also caused some issues with the unit tests which were simplified
for now.